pyproject.toml
[project]
name = "search-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.54.0",
    "fastmcp>=2.8.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "huggingface-hub>=0.33.1",
    "python-dotenv>=1.1.0",
]
